Bug 1493420 - Use a RwLock'd HashMap instead of a lock-free linked list for rule node children.
Summary:
I need to profile this a bit more, but talos was pretty happy about this, and it
solves the known performance issues here such as the test-case from bug 1483963
for example. This also gets rid of a bunch of unsafe code which is nice.

This still keeps the same GC scheme, removing the key from the hashmap when
needed. I kept those as release assertions, but should probably be turned into
debug-only assertions.

Test Plan:
No behavior change, so no new tests, though we should probably
consider landing a perf_reftest based on the testcase from bug 1483963...

Reviewers: bholley, heycam, xidorn

Tags: #secure-revision

Bug #: 1493420

Differential Revision: https://phabricator.services.mozilla.com/D6801